:truck: :: StageCardComponent -> GogoStawgeCardComponent
